requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
chardet==5.2.0
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.12
fpdf==1.7.2
greenlet==3.1.1
h11==0.14.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
pillow==11.2.1
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
python-dotenv==1.1.0
reportlab==4.3.1
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.1
